# v1/gvb_server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "API文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.01:9000",
    "basePath": "/",
    "paths": {
        "/api/advert/create": {
            "post": {
                "description": "创建广告",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "广告管理"
                ],
                "summary": "创建广告",
                "parameters": [
                    {
                        "description": "表示多个参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/advert_api.AdvertRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "advert_api.AdvertRequest": {
            "type": "object",
            "required": [
                "href",
                "images",
                "isShow",
                "title"
            ],
            "properties": {
                "href": {
                    "description": "跳转链接",
                    "type": "string"
                },
                "images": {
                    "description": "图片",
                    "type": "string"
                },
                "isShow": {
                    "description": "使用指针形式的原因:可以传入零值",
                    "type": "boolean"
                },
                "title": {
                    "description": "显示的标题",
                    "type": "string"
                }
            }
        },
...
        "res.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}